FEAT: New coloring logic for Results.
We are using Green to signal a good result.

We are using Dark Grey to signal a neutral result.
This is used for Inspection Type and Date.
(We should add logic in case the last inspection was old)

We are using Red to signal a bad result.

We are using Orange to signal an OK/Medium result.
We also use Orange with Violations to signal uncertainty:
1. It was shut down.
2. It was not ready to be inspected.

This is all handled by jQuery